feat: genesis double spend test
